+++ csatar-plugins/csatar/models/District.php
@@ -144,6 +144,14 @@ class District extends OrganizationBase
         ],
     ];
 
+    public $morphMany = [
+        'galleryPivot' => [
+            \Csatar\Csatar\Models\GalleryModelPivot::class,
+            'table' => 'csatar_csatar_gallery_model',
+            'name' => 'model',
+        ],
+    ];
+
     /**
      * Override the getExtendedNameAttribute function
      */
@@ -286,6 +294,14 @@ class District extends OrganizationBase
 
     public function getActiveMembersCountAttribute() {
         return StructureTree::getDistrictScoutsCount($this->id);
+    }
+
+    public function getTextForSearchResultsTreeAttribute() {
+        return $this->name;
+    }
+
+    public function getParentTree() {
+        return '(' . $this->association->text_for_search_results_tree . ')';
     }
 
 }