+++ csatar-plugins/csatar/models/Team.php
@@ -149,6 +149,14 @@ class Team extends OrganizationBase
      * Relations
      */
 
+    public $morphOne = [
+        'content_page' => [
+            '\Csatar\Csatar\Models\ContentPage',
+            'name' => 'model',
+            'label' => 'csatar.csatar::lang.plugin.admin.general.contentPage',
+        ],
+    ];
+
     public $belongsTo = [
         'district' => [
             '\Csatar\Csatar\Models\District',
@@ -156,6 +164,14 @@ class Team extends OrganizationBase
                 'requiredBeforeRender' => true,
             ],
             'label' => 'csatar.csatar::lang.plugin.admin.district.district',
+        ],
+    ];
+
+    public $morphMany = [
+        'galleryPivot' => [
+            \Csatar\Csatar\Models\GalleryModelPivot::class,
+            'table' => 'csatar_csatar_gallery_model',
+            'name' => 'model',
         ],
     ];
 
@@ -211,8 +227,12 @@ class Team extends OrganizationBase
             'ignoreInPermissionsMatrix' => true,
         ],
         'workPlans' => [
-            '\Csatar\Csatar\KnowledgeRepository\WorkPlan',
+            '\Csatar\KnowledgeRepository\Models\WorkPlan',
             'label' => 'csatar.knowledgerepository::lang.plugin.admin.workPlan.workPlans',
+        ],
+        'ovamtvWorkPlans' => [
+            '\Csatar\KnowledgeRepository\Models\OvamtvWorkPlan',
+            'label' => 'csatar.knowledgerepository::lang.plugin.admin.ovamtvWorkPlan.ovamtvWorkPlan',
         ],
     ];
 
@@ -349,14 +369,6 @@ class Team extends OrganizationBase
         return Team::find($id);
     }
 
-    public $morphOne = [
-        'content_page' => [
-            '\Csatar\Csatar\Models\ContentPage',
-            'name' => 'model',
-            'label' => 'csatar.csatar::lang.plugin.admin.general.contentPage',
-        ]
-    ];
-
     /**
      * Scope a query to only include teams with a given district id.
      */
@@ -468,6 +480,19 @@ class Team extends OrganizationBase
 
     public function getActiveMembersCountAttribute() {
         return StructureTree::getTeamScoutsCount($this->id);
+    }
+
+    public function getTextForSearchResultsTreeAttribute() {
+        return $this->team_number;
+    }
+
+    public function getParentTree() {
+        $tree = [
+            $this->district->association->text_for_search_results_tree,
+            $this->district->text_for_search_results_tree,
+        ];
+
+        return '(' . implode(' - ', $tree) . ')';
     }
 
 }